WIP: Omit the trampoline from non-rom code in gcl0x.py
This stops the trampoline from being written out into code that isn't
intended to be written to rom. This means that the full page is now
available to apps, but that the rom images written out by compilegcl are
now invalid.

Also add an infinite loop on the end of non-rom apps so that we don't just
start executing random code.