Add a hack to make emulator multi-page loading reliable
It seems I don't understand something either about data bus behaviour or
vsync/hsync timings, but the unreliability I think is down to incorrectly
timed data bus writing. Just add a hack that writes the whole byte to IN
immediately rather than a bit at a time. This would indicate that I'm
probably acting on hsync timing a bit late, but I don't know if that's
down to my error or some other inaccuracy in the emulation.

I'd like to figure out how to make this work reliably like the actual
hardware, but this will do for testing.