Add loading support to the emulator
A string, 'loader_file' is added to the top of the script to specify the
path of a .gt1 file. When the user presses 'L', the emulator will act like
an Arduino uploading code to the gigatron. Note, you must navigate to the
loader page manually first.

There are a couple of timing hacks in this code that I'd like to
understand and fix at some point, but this behaviour ought to be
quite close to the real hardware.